requirements.txt
# Core Requirements
flask==2.3.2
werkzeug==2.3.7

# Data Processing & ML
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
xgboost==1.7.5
joblib==1.2.0

# Chemistry & Visualization
rdkit-pypi==2022.9.5
plotly==5.15.0
shap==0.42.1

# Deployment & Environment
gunicorn==20.1.0
python-dotenv==1.0.0

# Additional Utilities
click==8.1.3
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3

flask-babel
flask-dance
fpdf